requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
pycparser==2.22
pycryptodome==3.21.0
setuptools==75.6.0
sqlparse==0.5.2
wheel==0.45.1
